+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres (more or less) to [Semantic Versioning](http://semver.o
 When you submit a PR, add your changes here!
 
 ### Fixed
+- context click actually calls double click callback #225
 - Removed href attribute from header divs #222
 
 ## 0.15.4